Random stuff

+++ src/Camera.c
@@ -82,9 +82,9 @@ void camera_processScrollWheel(Camera* camera, float yOffset){
 
 void updateCamera(Camera* camera){
     vec3_s front = vec3(0.0f,0.0f,0.0f);
-    front.x = cos(radians(camera->yaw)) * cos(radians(camera->pitch));
-    front.y = sin(radians(camera->pitch));
-    front.z = -(sin(radians(camera->yaw)) * cos(radians(camera->pitch)));
+    front.x = cos((double)radians(camera->yaw)) * cos((double)radians(camera->pitch));
+    front.y = sin((double)radians(camera->pitch));
+    front.z = -(sin((double)radians(camera->yaw)) * cos((double)radians(camera->pitch)));
 
 
     camera->front = vec3_normalize(front);
